chore: relocation integrations constants and update codeowners
