+++ packages/analytics-js-common/src/constants/integrations/config_to_integration_names.js
