+++ packages/analytics-js-common/src/constants/integrations/destinationNames.ts
@@ -1,316 +1,292 @@
 export {
   DISPLAY_NAME as AdobeAnalyticsDisplayName,
   DIR_NAME as AdobeAnalyticsDirectoryName,
-} from './integrations/AdobeAnalytics/constants';
+} from './AdobeAnalytics/constants';
 export {
   DISPLAY_NAME as AmplitudeDisplayName,
   DIR_NAME as AmplitudeDirectoryName,
-} from './integrations/Amplitude/constants';
+} from './Amplitude/constants';
 export {
   DISPLAY_NAME as AppcuesDisplayName,
   DIR_NAME as AppcuesDirectoryName,
-} from './integrations/Appcues/constants';
+} from './Appcues/constants';
 export {
   DISPLAY_NAME as BingAdsDisplayName,
   DIR_NAME as BingAdsDirectoryName,
-} from './integrations/BingAds/constants';
+} from './BingAds/constants';
 export {
   DISPLAY_NAME as BrazeDisplayName,
   DIR_NAME as BrazeDirectoryName,
-} from './integrations/Braze/constants';
+} from './Braze/constants';
 export {
   DISPLAY_NAME as BugsnagDisplayName,
   DIR_NAME as BugsnagDirectoryName,
-} from './integrations/Bugsnag/constants';
+} from './Bugsnag/constants';
 export {
   DISPLAY_NAME as ChartbeatDisplayName,
   DIR_NAME as ChartbeatDirectoryName,
-} from './integrations/Chartbeat/constants';
+} from './Chartbeat/constants';
 export {
   DISPLAY_NAME as CleverTapDisplayName,
   DIR_NAME as CleverTapDirectoryName,
-} from './integrations/Clevertap/constants';
+} from './Clevertap/constants';
 export {
   DISPLAY_NAME as ComscoreDisplayName,
   DIR_NAME as ComscoreDirectoryName,
-} from './integrations/Comscore/constants';
+} from './Comscore/constants';
 export {
   DISPLAY_NAME as CriteoDisplayName,
   DIR_NAME as CriteoDirectoryName,
-} from './integrations/Criteo/constants';
+} from './Criteo/constants';
 export {
   DISPLAY_NAME as CustomerIODisplayName,
   DIR_NAME as CustomerIODirectoryName,
-} from './integrations/CustomerIO/constants';
-export {
-  DISPLAY_NAME as DripDisplayName,
-  DIR_NAME as DripDirectoryName,
-} from './integrations/Drip/constants';
+} from './CustomerIO/constants';
+export { DISPLAY_NAME as DripDisplayName, DIR_NAME as DripDirectoryName } from './Drip/constants';
 export {
   DISPLAY_NAME as FacebookPixelDisplayName,
   DIR_NAME as FacebookPixelDirectoryName,
-} from './integrations/FacebookPixel/constants';
+} from './FacebookPixel/constants';
 export {
   DISPLAY_NAME as FullstoryDisplayName,
   DIR_NAME as FullstoryDirectoryName,
-} from './integrations/Fullstory/constants';
-export {
-  DISPLAY_NAME as GADisplayName,
-  DIR_NAME as GADirectoryName,
-} from './integrations/GA/constants';
-export {
-  DISPLAY_NAME as GA4DisplayName,
-  DIR_NAME as GA4DirectoryName,
-} from './integrations/GA4/constants';
+} from './Fullstory/constants';
+export { DISPLAY_NAME as GADisplayName, DIR_NAME as GADirectoryName } from './GA/constants';
+export { DISPLAY_NAME as GA4DisplayName, DIR_NAME as GA4DirectoryName } from './GA4/constants';
 export {
   DISPLAY_NAME as GA4_V2DisplayName,
   DIR_NAME as GA4_V2DirectoryName,
-} from './integrations/GA4_V2/constants';
+} from './GA4_V2/constants';
 export {
   DISPLAY_NAME as GoogleAdsDisplayName,
   DIR_NAME as GoogleAdsDirectoryName,
-} from './integrations/GoogleAds/constants';
+} from './GoogleAds/constants';
 export {
   DISPLAY_NAME as GoogleOptimizeDisplayName,
   DIR_NAME as GoogleOptimizeDirectoryName,
-} from './integrations/GoogleOptimize/constants';
+} from './GoogleOptimize/constants';
 export {
   DISPLAY_NAME as GoogleTagManagerDisplayName,
   DIR_NAME as GoogleTagManagerDirectoryName,
-} from './integrations/GoogleTagManager/constants';
-export {
-  DISPLAY_NAME as HeapDisplayName,
-  DIR_NAME as HeapDirectoryName,
-} from './integrations/Heap/constants';
+} from './GoogleTagManager/constants';
+export { DISPLAY_NAME as HeapDisplayName, DIR_NAME as HeapDirectoryName } from './Heap/constants';
 export {
   DISPLAY_NAME as HotjarDisplayName,
   DIR_NAME as HotjarDirectoryName,
-} from './integrations/Hotjar/constants';
+} from './Hotjar/constants';
 export {
   DISPLAY_NAME as HubSpotDisplayName,
   DIR_NAME as HubSpotDirectoryName,
-} from './integrations/HubSpot/constants';
+} from './HubSpot/constants';
 export {
   DISPLAY_NAME as IntercomDisplayName,
   DIR_NAME as IntercomDirectoryName,
-} from './integrations/INTERCOM/constants';
-export {
-  DISPLAY_NAME as KeenDisplayName,
-  DIR_NAME as KeenDirectoryName,
-} from './integrations/Keen/constants';
+} from './INTERCOM/constants';
+export { DISPLAY_NAME as KeenDisplayName, DIR_NAME as KeenDirectoryName } from './Keen/constants';
 export {
   DISPLAY_NAME as KissmetricsDisplayName,
   DIR_NAME as KissmetricsDirectoryName,
-} from './integrations/Kissmetrics/constants';
+} from './Kissmetrics/constants';
 export {
   DISPLAY_NAME as KlaviyoDisplayName,
   DIR_NAME as KlaviyoDirectoryName,
-} from './integrations/Klaviyo/constants';
+} from './Klaviyo/constants';
 export {
   DISPLAY_NAME as LaunchDarklyDisplayName,
   DIR_NAME as LaunchDarklyDirectoryName,
-} from './integrations/LaunchDarkly/constants';
+} from './LaunchDarkly/constants';
 export {
   DISPLAY_NAME as LinkedInInsightTagDisplayName,
   DIR_NAME as LinkedInInsightTagDirectoryName,
-} from './integrations/LinkedInInsightTag/constants';
+} from './LinkedInInsightTag/constants';
 export {
   DISPLAY_NAME as LotameDisplayName,
   DIR_NAME as LotameDirectoryName,
-} from './integrations/Lotame/constants';
+} from './Lotame/constants';
 export {
   DISPLAY_NAME as LyticsDisplayName,
   DIR_NAME as LyticsDirectoryName,
-} from './integrations/Lytics/constants';
+} from './Lytics/constants';
 export {
   DISPLAY_NAME as MixpanelDisplayName,
   DIR_NAME as MixpanelDirectoryName,
-} from './integrations/Mixpanel/constants';
+} from './Mixpanel/constants';
 export {
   DISPLAY_NAME as MoEngageDisplayName,
   DIR_NAME as MoEngageDirectoryName,
-} from './integrations/MoEngage/constants';
+} from './MoEngage/constants';
 export {
   DISPLAY_NAME as OptimizelyDisplayName,
   DIR_NAME as OptimizelyDirectoryName,
-} from './integrations/Optimizely/constants';
+} from './Optimizely/constants';
 export {
   DISPLAY_NAME as PendoDisplayName,
   DIR_NAME as PendoDirectoryName,
-} from './integrations/Pendo/constants';
+} from './Pendo/constants';
 export {
   DISPLAY_NAME as PinterestTagDisplayName,
   DIR_NAME as PinterestTagDirectoryName,
-} from './integrations/PinterestTag/constants';
+} from './PinterestTag/constants';
 export {
   DISPLAY_NAME as PostAffiliateProDisplayName,
   DIR_NAME as PostAffiliateProDirectoryName,
-} from './integrations/PostAffiliatePro/constants';
+} from './PostAffiliatePro/constants';
 export {
   DISPLAY_NAME as PostHogDisplayName,
   DIR_NAME as PostHogDirectoryName,
-} from './integrations/Posthog/constants';
+} from './Posthog/constants';
 export {
   DISPLAY_NAME as ProfitWellDisplayName,
   DIR_NAME as ProfitWellDirectoryName,
-} from './integrations/ProfitWell/constants';
+} from './ProfitWell/constants';
 export {
   DISPLAY_NAME as QualtricsDisplayName,
   DIR_NAME as QualtricsDirectoryName,
-} from './integrations/Qualtrics/constants';
+} from './Qualtrics/constants';
 export {
   DISPLAY_NAME as QuantumMetricDisplayName,
   DIR_NAME as QuantumMetricDirectoryName,
-} from './integrations/QuantumMetric/constants';
+} from './QuantumMetric/constants';
 export {
   DISPLAY_NAME as RedditPixelDisplayName,
   DIR_NAME as RedditPixelDirectoryName,
-} from './integrations/RedditPixel/constants';
+} from './RedditPixel/constants';
 export {
   DISPLAY_NAME as SentryDisplayName,
   DIR_NAME as SentryDirectoryName,
-} from './integrations/Sentry/constants';
+} from './Sentry/constants';
 export {
   DISPLAY_NAME as SnapPixelDisplayName,
   DIR_NAME as SnapPixelDirectoryName,
-} from './integrations/SnapPixel/constants';
+} from './SnapPixel/constants';
 export {
   DISPLAY_NAME as TVSquaredDisplayName,
   DIR_NAME as TVSquaredDirectoryName,
-} from './integrations/TVSquared/constants';
-export {
-  DISPLAY_NAME as VWODisplayName,
-  DIR_NAME as VWODirectoryName,
-} from './integrations/VWO/constants';
+} from './TVSquared/constants';
+export { DISPLAY_NAME as VWODisplayName, DIR_NAME as VWODirectoryName } from './VWO/constants';
 export {
   DISPLAY_NAME as GA360DisplayName,
   DIR_NAME as GA360DirectoryName,
-} from './integrations/GA360/constants';
+} from './GA360/constants';
 export {
   DISPLAY_NAME as AdrollDisplayName,
   DIR_NAME as AdrollDirectoryName,
-} from './integrations/Adroll/constants';
+} from './Adroll/constants';
 export {
   DISPLAY_NAME as DCMFloodlightDisplayName,
   DIR_NAME as DCMFloodlightDirectoryName,
-} from './integrations/DCMFloodlight/constants';
+} from './DCMFloodlight/constants';
 export {
   DISPLAY_NAME as MatomoDisplayName,
   DIR_NAME as MatomoDirectoryName,
-} from './integrations/Matomo/constants';
-export {
-  DISPLAY_NAME as VeroDisplayName,
-  DIR_NAME as VeroDirectoryName,
-} from './integrations/Vero/constants';
+} from './Matomo/constants';
+export { DISPLAY_NAME as VeroDisplayName, DIR_NAME as VeroDirectoryName } from './Vero/constants';
 export {
   DISPLAY_NAME as MouseflowDisplayName,
   DIR_NAME as MouseflowDirectoryName,
-} from './integrations/Mouseflow/constants';
+} from './Mouseflow/constants';
 export {
   DISPLAY_NAME as RockerboxDisplayName,
   DIR_NAME as RockerboxDirectoryName,
-} from './integrations/Rockerbox/constants';
+} from './Rockerbox/constants';
 export {
   DISPLAY_NAME as ConvertFlowDisplayName,
   DIR_NAME as ConvertFlowDirectoryName,
-} from './integrations/ConvertFlow/constants';
+} from './ConvertFlow/constants';
 export {
   DISPLAY_NAME as SnapEngageDisplayName,
   DIR_NAME as SnapEngageDirectoryName,
-} from './integrations/SnapEngage/constants';
+} from './SnapEngage/constants';
 export {
   DISPLAY_NAME as LiveChatDisplayName,
   DIR_NAME as LiveChatDirectoryName,
-} from './integrations/LiveChat/constants';
+} from './LiveChat/constants';
 export {
   DISPLAY_NAME as ShynetDisplayName,
   DIR_NAME as ShynetDirectoryName,
-} from './integrations/Shynet/constants';
+} from './Shynet/constants';
 export {
   DISPLAY_NAME as WoopraDisplayName,
   DIR_NAME as WoopraDirectoryName,
-} from './integrations/Woopra/constants';
+} from './Woopra/constants';
 export {
   DISPLAY_NAME as RollBarDisplayName,
   DIR_NAME as RollBarDirectoryName,
-} from './integrations/RollBar/constants';
+} from './RollBar/constants';
 export {
   DISPLAY_NAME as QuoraPixelDisplayName,
   DIR_NAME as QuoraPixelDirectoryName,
-} from './integrations/QuoraPixel/constants';
-export {
-  DISPLAY_NAME as JuneDisplayName,
-  DIR_NAME as JuneDirectoryName,
-} from './integrations/June/constants';
+} from './QuoraPixel/constants';
+export { DISPLAY_NAME as JuneDisplayName, DIR_NAME as JuneDirectoryName } from './June/constants';
 export {
   DISPLAY_NAME as EngageDisplayName,
   DIR_NAME as EngageDirectoryName,
-} from './integrations/Engage/constants';
+} from './Engage/constants';
 export {
   DISPLAY_NAME as IterableDisplayName,
   DIR_NAME as IterableDirectoryName,
-} from './integrations/Iterable/constants';
+} from './Iterable/constants';
 export {
   DISPLAY_NAME as YandexMetricaDisplayName,
   DIR_NAME as YandexMetricaDirectoryName,
-} from './integrations/YandexMetrica/constants';
+} from './YandexMetrica/constants';
 export {
   DISPLAY_NAME as RefinerDisplayName,
   DIR_NAME as RefinerDirectoryName,
-} from './integrations/Refiner/constants';
+} from './Refiner/constants';
 export {
   DISPLAY_NAME as QualarooDisplayName,
   DIR_NAME as QualarooDirectoryName,
-} from './integrations/Qualaroo/constants';
+} from './Qualaroo/constants';
 export {
   DISPLAY_NAME as PodsightsDisplayName,
   DIR_NAME as PodsightsDirectoryName,
-} from './integrations/Podsights/constants';
+} from './Podsights/constants';
 export {
   DISPLAY_NAME as AxeptioDisplayName,
   DIR_NAME as AxeptioDirectoryName,
-} from './integrations/Axeptio/constants';
+} from './Axeptio/constants';
 export {
   DISPLAY_NAME as SatismeterDisplayName,
   DIR_NAME as SatismeterDirectoryName,
-} from './integrations/Satismeter/constants';
+} from './Satismeter/constants';
 export {
   DISPLAY_NAME as MicrosoftClarityDisplayName,
   DIR_NAME as MicrosoftClarityDirectoryName,
-} from './integrations/MicrosoftClarity/constants';
+} from './MicrosoftClarity/constants';
 export {
   DISPLAY_NAME as SendinblueDisplayName,
   DIR_NAME as SendinblueDirectoryName,
-} from './integrations/Sendinblue/constants';
+} from './Sendinblue/constants';
 export {
   DISPLAY_NAME as OlarkDisplayName,
   DIR_NAME as OlarkDirectoryName,
-} from './integrations/Olark/constants';
+} from './Olark/constants';
 export {
   DISPLAY_NAME as LemniskDisplayName,
   DIR_NAME as LemniskDirectoryName,
-} from './integrations/Lemnisk/constants';
+} from './Lemnisk/constants';
 export {
   DISPLAY_NAME as TiktokAdsDisplayName,
   DIR_NAME as TiktokAdsDirectoryName,
-} from './integrations/TiktokAds/constants';
+} from './TiktokAds/constants';
 export {
   DISPLAY_NAME as ActiveCampaignDisplayName,
   DIR_NAME as ActiveCampaignDirectoryName,
-} from './integrations/ActiveCampaign/constants';
+} from './ActiveCampaign/constants';
 export {
   DISPLAY_NAME as SprigDisplayName,
   DIR_NAME as SprigDirectoryName,
-} from './integrations/Sprig/constants';
+} from './Sprig/constants';
 export {
   DISPLAY_NAME as SpotifyPixelDisplayName,
   DIR_NAME as SpotifyPixelDirectoryName,
-} from './integrations/SpotifyPixel/constants';
+} from './SpotifyPixel/constants';
 export {
   DISPLAY_NAME as CommandBarDisplayName,
   DIR_NAME as CommandBarDirectoryName,
-} from './integrations/CommandBar/constants';
+} from './CommandBar/constants';
 export {
   DISPLAY_NAME as NinetailedDisplayName,
   DIR_NAME as NinetailedDirectoryName,
-} from './integrations/Ninetailed/constants';
+} from './Ninetailed/constants';
