+++ packages/analytics-js-common/src/v1.1/utils/index.ts
@@ -1,9 +1,6 @@
 export * as storage from './storage';
-export * as client_server_name from './client_server_name';
-export * as config_to_integration_names from './config_to_integration_names';
 export * as constants from './constants';
 export * as errorHandler from './errorHandler';
-export * as integration_cname from './integration_cname';
 export * as logUtil from './logUtil';
 export * as notifyError from './notifyError';
 export * as ObjectUtils from './ObjectUtils';
