--- packages/analytics-js-common/src/v1.1/utils/integration_cname.js
@@ -1,166 +1,0 @@
-import { CNameMapping as AdobeAnalytics } from '../../constants/integrations/AdobeAnalytics/constants';
-import { CNameMapping as Amplitude } from '../../constants/integrations/Amplitude/constants';
-import { CNameMapping as Appcues } from '../../constants/integrations/Appcues/constants';
-import { CNameMapping as BingAds } from '../../constants/integrations/BingAds/constants';
-import { CNameMapping as Braze } from '../../constants/integrations/Braze/constants';
-import { CNameMapping as Bugsnag } from '../../constants/integrations/Bugsnag/constants';
-import { CNameMapping as Chartbeat } from '../../constants/integrations/Chartbeat/constants';
-import { CNameMapping as Clevertap } from '../../constants/integrations/Clevertap/constants';
-import { CNameMapping as Comscore } from '../../constants/integrations/Comscore/constants';
-import { CNameMapping as Criteo } from '../../constants/integrations/Criteo/constants';
-import { CNameMapping as CustomerIO } from '../../constants/integrations/CustomerIO/constants';
-import { CNameMapping as Drip } from '../../constants/integrations/Drip/constants';
-import { CNameMapping as FacebookPixel } from '../../constants/integrations/FacebookPixel/constants';
-import { CNameMapping as Fullstory } from '../../constants/integrations/Fullstory/constants';
-import { CNameMapping as GA } from '../../constants/integrations/GA/constants';
-import { CNameMapping as GA4 } from '../../constants/integrations/GA4/constants';
-import { CNameMapping as GA4_V2 } from '../../constants/integrations/GA4_V2/constants';
-import { CNameMapping as GoogleAds } from '../../constants/integrations/GoogleAds/constants';
-import { CNameMapping as GoogleOptimize } from '../../constants/integrations/GoogleOptimize/constants';
-import { CNameMapping as GoogleTagManager } from '../../constants/integrations/GoogleTagManager/constants';
-import { CNameMapping as Heap } from '../../constants/integrations/Heap/constants';
-import { CNameMapping as Hotjar } from '../../constants/integrations/Hotjar/constants';
-import { CNameMapping as HubSpot } from '../../constants/integrations/HubSpot/constants';
-import { CNameMapping as INTERCOM } from '../../constants/integrations/INTERCOM/constants';
-import { CNameMapping as Keen } from '../../constants/integrations/Keen/constants';
-import { CNameMapping as Kissmetrics } from '../../constants/integrations/Kissmetrics/constants';
-import { CNameMapping as Klaviyo } from '../../constants/integrations/Klaviyo/constants';
-import { CNameMapping as LaunchDarkly } from '../../constants/integrations/LaunchDarkly/constants';
-import { CNameMapping as LinkedInInsightTag } from '../../constants/integrations/LinkedInInsightTag/constants';
-import { CNameMapping as Lotame } from '../../constants/integrations/Lotame/constants';
-import { CNameMapping as Lytics } from '../../constants/integrations/Lytics/constants';
-import { CNameMapping as Mixpanel } from '../../constants/integrations/Mixpanel/constants';
-import { CNameMapping as MoEngage } from '../../constants/integrations/MoEngage/constants';
-import { CNameMapping as Optimizely } from '../../constants/integrations/Optimizely/constants';
-import { CNameMapping as Pendo } from '../../constants/integrations/Pendo/constants';
-import { CNameMapping as PinterestTag } from '../../constants/integrations/PinterestTag/constants';
-import { CNameMapping as PostAffiliatePro } from '../../constants/integrations/PostAffiliatePro/constants';
-import { CNameMapping as Posthog } from '../../constants/integrations/Posthog/constants';
-import { CNameMapping as ProfitWell } from '../../constants/integrations/ProfitWell/constants';
-import { CNameMapping as Qualtrics } from '../../constants/integrations/Qualtrics/constants';
-import { CNameMapping as QuantumMetric } from '../../constants/integrations/QuantumMetric/constants';
-import { CNameMapping as RedditPixel } from '../../constants/integrations/RedditPixel/constants';
-import { CNameMapping as Sentry } from '../../constants/integrations/Sentry/constants';
-import { CNameMapping as SnapPixel } from '../../constants/integrations/SnapPixel/constants';
-import { CNameMapping as TVSquared } from '../../constants/integrations/TVSquared/constants';
-import { CNameMapping as VWO } from '../../constants/integrations/VWO/constants';
-import { CNameMapping as GA360 } from '../../constants/integrations/GA360/constants';
-import { CNameMapping as Adroll } from '../../constants/integrations/Adroll/constants';
-import { CNameMapping as DCMFloodlight } from '../../constants/integrations/DCMFloodlight/constants';
-import { CNameMapping as Matomo } from '../../constants/integrations/Matomo/constants';
-import { CNameMapping as Vero } from '../../constants/integrations/Vero/constants';
-import { CNameMapping as Mouseflow } from '../../constants/integrations/Mouseflow/constants';
-import { CNameMapping as Rockerbox } from '../../constants/integrations/Rockerbox/constants';
-import { CNameMapping as ConvertFlow } from '../../constants/integrations/ConvertFlow/constants';
-import { CNameMapping as SnapEngage } from '../../constants/integrations/SnapEngage/constants';
-import { CNameMapping as LiveChat } from '../../constants/integrations/LiveChat/constants';
-import { CNameMapping as Shynet } from '../../constants/integrations/Shynet/constants';
-import { CNameMapping as Woopra } from '../../constants/integrations/Woopra/constants';
-import { CNameMapping as RollBar } from '../../constants/integrations/RollBar/constants';
-import { CNameMapping as QuoraPixel } from '../../constants/integrations/QuoraPixel/constants';
-import { CNameMapping as June } from '../../constants/integrations/June/constants';
-import { CNameMapping as Engage } from '../../constants/integrations/Engage/constants';
-import { CNameMapping as Iterable } from '../../constants/integrations/Iterable/constants';
-import { CNameMapping as YandexMetrica } from '../../constants/integrations/YandexMetrica/constants';
-import { CNameMapping as Refiner } from '../../constants/integrations/Refiner/constants';
-import { CNameMapping as Qualaroo } from '../../constants/integrations/Qualaroo/constants';
-import { CNameMapping as Podsights } from '../../constants/integrations/Podsights/constants';
-import { CNameMapping as Axeptio } from '../../constants/integrations/Axeptio/constants';
-import { CNameMapping as SatisMeter } from '../../constants/integrations/Satismeter/constants';
-import { CNameMapping as MicrosoftClarity } from '../../constants/integrations/MicrosoftClarity/constants';
-import { CNameMapping as Sendinblue } from '../../constants/integrations/Sendinblue/constants';
-import { CNameMapping as Olark } from '../../constants/integrations/Olark/constants';
-import { CNameMapping as Lemnisk } from '../../constants/integrations/Lemnisk/constants';
-import { CNameMapping as TiktokAds } from '../../constants/integrations/TiktokAds/constants';
-import { CNameMapping as ActiveCampaign } from '../../constants/integrations/ActiveCampaign/constants';
-import { CNameMapping as Sprig } from '../../constants/integrations/Sprig/constants';
-import { CNameMapping as SpotifyPixel } from '../../constants/integrations/SpotifyPixel/constants';
-import { CNameMapping as CommandBar } from '../../constants/integrations/CommandBar/constants';
-import { CNameMapping as Ninetailed } from '../../constants/integrations/Ninetailed/constants';
-
-// for sdk side native integration identification
-// add a mapping from common names to index.js exported key names as identified by Rudder
-const commonNames = {
-  All: 'All',
-  ...AdobeAnalytics,
-  ...Amplitude,
-  ...Appcues,
-  ...BingAds,
-  ...Braze,
-  ...Bugsnag,
-  ...CommandBar,
-  ...Chartbeat,
-  ...Clevertap,
-  ...Comscore,
-  ...Criteo,
-  ...CustomerIO,
-  ...Drip,
-  ...FacebookPixel,
-  ...Fullstory,
-  ...GA,
-  ...GA4,
-  ...GA4_V2,
-  ...GA360,
-  ...GoogleAds,
-  ...GoogleOptimize,
-  ...GoogleTagManager,
-  ...Heap,
-  ...Hotjar,
-  ...HubSpot,
-  ...INTERCOM,
-  ...Keen,
-  ...Kissmetrics,
-  ...Klaviyo,
-  ...LaunchDarkly,
-  ...LinkedInInsightTag,
-  ...Lotame,
-  ...Lytics,
-  ...Mixpanel,
-  ...MoEngage,
-  ...Ninetailed,
-  ...Optimizely,
-  ...Pendo,
-  ...PinterestTag,
-  ...PostAffiliatePro,
-  ...Posthog,
-  ...ProfitWell,
-  ...Qualtrics,
-  ...QuantumMetric,
-  ...RedditPixel,
-  ...Sentry,
-  ...SnapPixel,
-  ...TVSquared,
-  ...VWO,
-  ...Adroll,
-  ...DCMFloodlight,
-  ...Matomo,
-  ...Vero,
-  ...Mouseflow,
-  ...ConvertFlow,
-  ...SnapEngage,
-  ...LiveChat,
-  ...Shynet,
-  ...Woopra,
-  ...RollBar,
-  ...QuoraPixel,
-  ...June,
-  ...Engage,
-  ...Iterable,
-  ...Rockerbox,
-  ...YandexMetrica,
-  ...Refiner,
-  ...Qualaroo,
-  ...Podsights,
-  ...Axeptio,
-  ...SatisMeter,
-  ...MicrosoftClarity,
-  ...Sendinblue,
-  ...Olark,
-  ...Lemnisk,
-  ...TiktokAds,
-  ...ActiveCampaign,
-  ...Sprig,
-  ...SpotifyPixel,
-};
-
-export { commonNames };
